backend/requirements.txt
# Backend API Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Agent_Tools dependencies (from requirements_analysis.txt)
gseapy>=1.0.0
statsmodels>=0.13.0
matplotlib>=3.5.0
seaborn>=0.12.0
networkx>=2.8.0
goatools>=1.3.0
scipy>=1.9.0
pandas>=1.5.0
numpy>=1.23.0
anndata>=0.8.0
scanpy>=1.9.0

# Additional dependencies for pipeline
torch>=2.0.0
scikit-learn>=1.0.0

